회원 탈퇴시 연관 데이터 유지

+++ src/main/java/com/panda/back/domain/member/service/MemberService.java
@@ -1,5 +1,7 @@
 package com.panda.back.domain.member.service;
 
+import com.panda.back.domain.item.entity.Item;
+import com.panda.back.domain.item.repository.ItemRepository;
 import com.panda.back.domain.member.dto.PasswordRequestDto;
 import com.panda.back.domain.member.dto.ProfileRequestDto;
 import com.panda.back.domain.member.dto.ProfileResponseDto;
@@ -21,6 +23,7 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -31,16 +34,17 @@ public class MemberService {
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
     private final S3Uploader s3Uploader;
+    private final ItemRepository itemRepository;
 
     public BaseResponse membernameExists(String membername) {
-        if (memberRepository.findByMembername(membername).isPresent()){
+        if (memberRepository.findByMembername(membername).isPresent()) {
             throw new CustomException(ErrorCode.DUPLICATE_MEMBERNAME);
         }
         return BaseResponse.successMessage("중복 체크 완료");
     }
 
     public BaseResponse nicknameExists(String nickname) {
-        if (memberRepository.findByNickname(nickname).isPresent()){
+        if (memberRepository.findByNickname(nickname).isPresent()) {
             throw new CustomException(ErrorCode.DUPLICATE_NICKNAME);
         }
         return BaseResponse.successMessage("중복 체크 완료");
@@ -85,21 +89,21 @@ public class MemberService {
     }
 
     @Transactional
-    public BaseResponse updateProfile(ProfileRequestDto requestDto, Member member) {
-        try {
-            // 현재 로그인한 사용자의 정보를 가져옴
-            Member myprofile = findByMembername(member.getMembername());
-            myprofile.setNickname(requestDto.getNickname());
-            myprofile.setIntro(requestDto.getIntro());
-            memberRepository.save(myprofile);
-            return BaseResponse.successMessage("회원정보 수정 성공");
-        } catch (IllegalArgumentException e) {
-            return BaseResponse.error(e.getMessage());
+    public ResponseEntity<BaseResponse<String>> updateProfile(ProfileRequestDto requestDto, Member member, BindingResult bindingResult) {
+        List<FieldError> fieldErrors = bindingResult.getFieldErrors();
+        for (FieldError e : fieldErrors) {
+            return ResponseEntity.status(HttpStatus.NOT_ACCEPTABLE).body(BaseResponse.error(e.getDefaultMessage()));
         }
+        // 현재 로그인한 사용자의 정보를 가져옴
+        Member myprofile = findByMembername(member.getMembername());
+        myprofile.setNickname(requestDto.getNickname());
+        myprofile.setIntro(requestDto.getIntro());
+        memberRepository.save(myprofile);
+        return ResponseEntity.status(HttpStatus.OK).body(BaseResponse.successMessage("회원정보 수정 성공"));
     }
 
     @Transactional
-    public BaseResponse updatePassword(PasswordRequestDto requestDto, Member member){
+    public BaseResponse updatePassword(PasswordRequestDto requestDto, Member member) {
         try {
             Member myprofile = findByMembername(member.getMembername());
 
@@ -120,19 +124,31 @@ public class MemberService {
 
     @Transactional
     public BaseResponse deleteMember(Member member) {
-       Member currentMember = findByMembername(member.getMembername());
-        if(!Objects.equals(member.getProfileImageUrl(), "https://bidpanda-bucket.s3.ap-northeast-2.amazonaws.com/defualt-image/IMG_0191.png")){
+        Member currentMember = findByMembername(member.getMembername());
+
+        //멤버가 등록한 아이템 중 AuctionEndTime이 현재시간보다 이후이면 삭제 불가
+        if (itemRepository.existsByMemberAndAuctionEndTimeAfter(currentMember, LocalDateTime.now())) {
+            throw new CustomException(ErrorCode.EXIST_IS_NOT_CLOSED_ITEMS);
+        }
+
+        if (!Objects.equals(member.getProfileImageUrl(), "https://bidpanda-bucket.s3.ap-northeast-2.amazonaws.com/defualt-image/IMG_0191.png")) {
             String fileName = member.getProfileImageUrl().substring(member.getProfileImageUrl().lastIndexOf("com") + 4);
             s3Uploader.deleteFile(fileName);
         }
         memberRepository.delete(currentMember);
+
+        List<Item> memberItems = itemRepository.findAllByMember(member);
+        for (Item item : memberItems) {
+            item.setMember(null);
+            itemRepository.save(item);
+        }
         return BaseResponse.successMessage("회원 삭제 성공");
     }
 
     @Transactional
     public BaseResponse updateProfileImage(MultipartFile file, Member member) throws IOException {
         String url = s3Uploader.upload(file, "profile");
-        if(!Objects.equals(member.getProfileImageUrl(), "https://bidpanda-bucket.s3.ap-northeast-2.amazonaws.com/defualt-image/IMG_0191.png")){
+        if (!Objects.equals(member.getProfileImageUrl(), "https://bidpanda-bucket.s3.ap-northeast-2.amazonaws.com/defualt-image/IMG_0191.png")) {
             String fileName = member.getProfileImageUrl().substring(member.getProfileImageUrl().lastIndexOf("com") + 4);
             s3Uploader.deleteFile(fileName);
         }
@@ -140,8 +156,8 @@ public class MemberService {
         memberRepository.save(member);
         return BaseResponse.successData(url);
     }
-  
-  
+
+
     public Member findByMembername(String membername) {
         return memberRepository.findByMembername(membername).orElseThrow(() ->
                 new CustomException(ErrorCode.NOT_FOUND_MEMBER));
